first/requirements.txt
Django==2.0.3
numpy
Pillow
pytz
scikit-learn
scipy
